Loading Spinner moves to be reusable + Moves timing icons to own file
